Fix missing version info from Git Deploy
In update.php, $module_list is a global variable that Git Deploy can add itself to. In the updatedb command it is a local variable. This causes every module that has a version dependency to generate errors. The fix is to make $module_list a global variable as in update.php.